Moving invalid_utf8 tests into a separate mod
