+++ parquet/tests/arrow_reader/mod.rs
@@ -43,6 +43,7 @@ mod bloom_filter;
 #[cfg(feature = "crc")]
 mod checksum;
 mod int96_stats_roundtrip;
+mod invalid_utf8;
 mod io;
 #[cfg(feature = "async")]
 mod predicate_cache;
